move query checks into own files
